Split logic for 2 modules
