+++ build.gradle
@@ -1,6 +1,8 @@
 // Top-level build file where you can add configuration options common to all sub-projects/modules.
 
-buildscript {
+buildscript { 
+    ext.kotlin_version = '1.3.50'
+
     apply from: 'versions.gradle'
     apply from: "secrets.gradle"
