--- visionai-core/src/main/java/com/bendenen/visionai/utils/NativeImageUtilsWrapper.kt
@@ -1,26 +1,0 @@
-package com.bendenen.visionai.utils
-
-object NativeImageUtilsWrapper {
-
-        init {
-            System.loadLibrary("image-utils")
-        }
-
-    external fun resizeImage(
-        imageByteArray: ByteArray,
-        imageWidth: Int,
-        imageHeight: Int,
-        newWidth: Int,
-        newHeight: Int,
-        outputArray: ByteArray
-    )
-
-    external fun resizeAndNormalizeImage(
-        imageByteArray: ByteArray,
-        imageWidth: Int,
-        imageHeight: Int,
-        newWidth: Int,
-        newHeight: Int,
-        outputArray: FloatArray
-    )
-}
