+++ visionai-core/src/main/java/com/bendenen/visionai/videoprocessor/VideoProcessorListener.kt
@@ -5,4 +5,6 @@ import android.graphics.Bitmap
 interface VideoProcessorListener {
 
     fun onNewFrameProcessed(bitmap: Bitmap)
+
+    fun onFinish()
 }
