+++ visionai-tflite-processor/.gitignore
@@ -1,0 +1,7 @@
+/build
+*~
+.externalNativeBuild
+libwebp
+.DS_Store
+**/ndkHelperBin
+**/.cxx
