+++ visionai-tflite-processor/src/main/assets/detect.tflite
