+++ visionai-tflite-processor/src/main/assets/labelmap.txt
