+++ visionai-tflite-processor/src/main/assets/test_style.jpeg
