+++ visionai-tflite-processor/src/main/cpp/utils/image_utils.hpp
