+++ visionai-tflite-processor/src/main/java/com/bendenen/visionai/tflite/utils/NativeImageUtilsWrapper.kt
@@ -1,0 +1,26 @@
+package com.bendenen.visionai.tflite.utils
+
+object NativeImageUtilsWrapper {
+
+        init {
+            System.loadLibrary("image-utils")
+        }
+
+    external fun resizeImage(
+        imageByteArray: ByteArray,
+        imageWidth: Int,
+        imageHeight: Int,
+        newWidth: Int,
+        newHeight: Int,
+        outputArray: ByteArray
+    )
+
+    external fun resizeAndNormalizeImage(
+        imageByteArray: ByteArray,
+        imageWidth: Int,
+        imageHeight: Int,
+        newWidth: Int,
+        newHeight: Int,
+        outputArray: FloatArray
+    )
+}
